Scripts are ready to be merged.
